Task 2

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -13,11 +13,11 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned long i;
+	int i;
 
 	if (array != NULL || action != NULL)
 	{
-		for (i = 0; i < size; i++)
+		for (i = 0; i < (int) size; i++)
 		{
 			action(array[i]);
 		}
